type of TaskHandle is void* not uInt32
at least in NIDAQmx 15.1 TaskHandle is defined as a void* and not as a
uInt32. In practice this makes no difference on 32 bit system (or 64bit
systems running 32bit Python), however with 64bit bit it does.

This is tested on a Windows 7 64-bit using 64bit WinPythin. The system is
running a set of 6353 cards currently using CounterOutputs and
AnalogOutputs. This is not tested on 32 bit python, but it shouldn't have
any changes on these systems.