don't call .data on numpy arrays for ctypes
calling data on a ctypes array gives the address as a number, not as a
pointer type. This is fine for pointer addresses < 2**32-1. Larger
addresses cause conversion errors as it tries to fit it into a 32 bit
integer.

The solution is to remove the call to data. This passes a ctypes object
to the function which calls _as_paramater_ on that object. That returns
the address in a ctypes.c_void_p which is of the correct pointer size.