Python2/Python3 compatibility
This patch changes interactions between the c-library to use
.encode/.decode when using strings.  This is necessary for Python3 and
should generally be compatible for Python2.

This patch also changes the DAQmxConstants class to not be a namedtuple,
but rather a (mostly) immutable dict specialization.  This was necessary
because many versions of Python limit the number of function arguments to
255.  This change keeps DAQmx.* items (mostly) immutable and also
accessible using a pattern like DAQmx.value (rather than DAQmx['value']).

Otherwise, this patch changes print statements, and includes more
'from __future__ import print_function' statements.  This patch also uses
the six.moves module to import input(py3)/raw_input(py2) appropriately for
some of the examples to ensure compatibility between Python2 and Python3.
The six module is not used for the main library.

This patch did _not_ need to do anything with repsect to changing division
operations, since that had been done with earlier commits.

This patch does _not_ do anything to the packaging scripts or
documentation; this is expected to be done at a later date.